+++ changelog.md
@@ -16,6 +16,7 @@
 * [3750](https://github.com/zeta-chain/node/pull/3750) - support simple call from solana
 * [3764](https://github.com/zeta-chain/node/pull/3764) - add payload parsing for Sui WithdrawAndCall
 * [3756](https://github.com/zeta-chain/node/pull/3756) - parse revert options in solana inbounds
+* [3765](https://github.com/zeta-chain/node/pull/3765) - support cancelling Sui rejected withdrawal
 * [3790](https://github.com/zeta-chain/node/pull/3790) - integrate execute revert
 
 ### Refactor